+++ samples/sample.ts
@@ -1,7 +1,7 @@
 let x : boolean = true;
 let y : number = 5;
 let z : number = 8;
-let t = "ttt";
+let t = "ttttttttttttttttt";
 const a : string = "I love ";
                     const b : string = "Typescript";
                     const c : string = "I love " + "Typescript";
@@ -13,3 +13,9 @@ const a : string = "I love ";
         let z = y;
     }
 }
+const walid = "Happy birthday";
+const l = walid.substring(3);
+const hi = walid.substring(2);
+
+const f = d + s;
+2--;